Move SqlCommon Benchmark tests used for comparison to separate Project so we can skip in CI
