+++ Akka.Persistence.Linq2Db.Benchmark.DockerComparisonTests/SqlServerSpecsFixture.cs
@@ -1,0 +1,11 @@
+﻿using Akka.Persistence.Sql.Linq2Db.Tests.Docker;
+using Akka.Persistence.Sql.Linq2Db.Tests.Docker.Docker;
+using Xunit;
+
+namespace Akka.Persistence.Linq2Db.BenchmarkTests.Docker.Linq2Db
+{
+    [CollectionDefinition("SqlServerSpec")]
+    public sealed class SqlServerSpecsFixture : ICollectionFixture<SqlServerFixture>
+    {
+    }
+}
